requirement.txt
# Libraries required for the project
pymongo==4.3.3                  # MongoDB client for interacting with MongoDB
python-telegram-bot==20.1       # Python Telegram Bot library for interacting with Telegram
google-generativeai==0.0.3      # Google Gemini AI library for generating responses
requests==2.28.2                # Library for making HTTP requests for web search


# Instructions for setting up API keys:

# 1. **Gemini API Key**:
#    - Replace the placeholder `GEN_API_KEY` in the code with your Gemini API key.
#
# 2. **Web Search API Key**:
#    - Replace the placeholder `WEB_SEARCH_API_KEY` in the code with your Web Search API key.
#
# 3. **MongoDB Connection String**:
#    - Replace the placeholder connection string in the code with your own MongoDB credentials.
